move `From<Flags>` impl to cli/flags.rs
